+++ src/main/kotlin/com/github/fstaudt/aoc2023/day2/Day2.kt
@@ -22,7 +22,7 @@ class Day2(fileName: String = "day_2.txt") : Day {
 
     private fun String.toGame(): Game {
         return split(":").let { game ->
-            val id = Regex("Game (\\d+)").find(game[0])!!.groupValues[1].toInt()
+            val id = Regex("Game (\\d+)").find(game[0])!!.groupValues[1].toLong()
             val subsets = game[1].split(";").map { it.toSubset() }
             Game(id, subsets)
         }
@@ -34,7 +34,7 @@ class Day2(fileName: String = "day_2.txt") : Day {
 
     private fun String.toSingleColor(): Subset {
         return Regex(" *(\\d+) (red|green|blue)").find(this)!!.let {
-            val count = it.groupValues[1].toInt()
+            val count = it.groupValues[1].toLong()
             when (it.groupValues[2]) {
                 "red" -> Subset(count, 0, 0)
                 "green" -> Subset(0, count, 0)
@@ -43,12 +43,12 @@ class Day2(fileName: String = "day_2.txt") : Day {
         }
     }
 
-    data class Game(val id: Int, val subsets: List<Subset>) {
+    data class Game(val id: Long, val subsets: List<Subset>) {
         fun isValid() = subsets.all { it.red <= 12 && it.green <= 13 && it.blue <= 14 }
         fun toMinimalSet() = subsets.reduce { acc, subset -> acc.max(subset) }
     }
 
-    data class Subset(val red: Int, val green: Int, val blue: Int) {
+    data class Subset(val red: Long, val green: Long, val blue: Long) {
         fun plus(subset: Subset) = Subset(red + subset.red, green + subset.green, blue + subset.blue)
         fun max(subset: Subset) = Subset(max(red, subset.red), max(green, subset.green), max(blue, subset.blue))
         fun toPower() = red * green * blue
